% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger.R
\name{EdgeRClass}
\alias{EdgeRClass}
\title{R6 class for EdgeR analysis}
\value{

}
\description{
R6 class for EdgeR analysis

R6 class for EdgeR analysis
}
\examples{
edger <- EdgeRClass$new(salmon, groupcol, outdir)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EdgeRClass$new()}}
\item \href{#method-EdgerToLimma}{\code{EdgeRClass$EdgerToLimma()}}
\item \href{#method-runEdger}{\code{EdgeRClass$runEdger()}}
\item \href{#method-getResults}{\code{EdgeRClass$getResults()}}
\item \href{#method-clone}{\code{EdgeRClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdgeRClass$new(salmon, groupcol, outdir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{salmon}}{salmon count data}

\item{\code{groupcol}}{category to use for DGE contrasts}

\item{\code{outdir}}{directory to write DGE results to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EdgerToLimma"></a>}}
\if{latex}{\out{\hypertarget{method-EdgerToLimma}{}}}
\subsection{Method \code{EdgerToLimma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdgeRClass$EdgerToLimma(p = NULL, padj = 0.05, logfc = 1.5, outdir, groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outdir}}{directory to write DGE results to}

\item{\code{groupcol}}{category to use for DGE contrasts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-runEdger"></a>}}
\if{latex}{\out{\hypertarget{method-runEdger}{}}}
\subsection{Method \code{runEdger()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdgeRClass$runEdger(p = NULL, padj = 0.05, logfc = 1.5, outdir, groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outdir}}{directory to write DGE results to}

\item{\code{groupcol}}{category to use for DGE contrasts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResults"></a>}}
\if{latex}{\out{\hypertarget{method-getResults}{}}}
\subsection{Method \code{getResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdgeRClass$getResults(p = NULL, padj = 0.05, logfc = 1.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EdgeRClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
