% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{MixomicsClass}
\alias{MixomicsClass}
\title{R6 class for mixomics analysis}
\value{

}
\description{
R6 class for mixomics analysis

R6 class for mixomics analysis
}
\examples{
mixomixs <- MixomicsClass$new(result, datTraits)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MixomicsClass$new()}}
\item \href{#method-evaluatePerformance}{\code{MixomicsClass$evaluatePerformance()}}
\item \href{#method-plotPerformance}{\code{MixomicsClass$plotPerformance()}}
\item \href{#method-tune}{\code{MixomicsClass$tune()}}
\item \href{#method-getFinalModel}{\code{MixomicsClass$getFinalModel()}}
\item \href{#method-plotDiablo}{\code{MixomicsClass$plotDiablo()}}
\item \href{#method-plotIndiv}{\code{MixomicsClass$plotIndiv()}}
\item \href{#method-plotArrow}{\code{MixomicsClass$plotArrow()}}
\item \href{#method-plotVar}{\code{MixomicsClass$plotVar()}}
\item \href{#method-plotCircos}{\code{MixomicsClass$plotCircos()}}
\item \href{#method-plotLoadings}{\code{MixomicsClass$plotLoadings()}}
\item \href{#method-plotNetwork}{\code{MixomicsClass$plotNetwork()}}
\item \href{#method-plotHeatmap}{\code{MixomicsClass$plotHeatmap()}}
\item \href{#method-cytoscapeImportTable}{\code{MixomicsClass$cytoscapeImportTable()}}
\item \href{#method-clone}{\code{MixomicsClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$new(result, datTraits, genecode = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{DESeq2 analysis result}

\item{\code{datTraits}}{clinicaldata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluatePerformance"></a>}}
\if{latex}{\out{\hypertarget{method-evaluatePerformance}{}}}
\subsection{Method \code{evaluatePerformance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$evaluatePerformance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPerformance"></a>}}
\if{latex}{\out{\hypertarget{method-plotPerformance}{}}}
\subsection{Method \code{plotPerformance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotPerformance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tune"></a>}}
\if{latex}{\out{\hypertarget{method-tune}{}}}
\subsection{Method \code{tune()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$tune()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFinalModel"></a>}}
\if{latex}{\out{\hypertarget{method-getFinalModel}{}}}
\subsection{Method \code{getFinalModel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$getFinalModel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDiablo"></a>}}
\if{latex}{\out{\hypertarget{method-plotDiablo}{}}}
\subsection{Method \code{plotDiablo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotDiablo(ncomp = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotIndiv"></a>}}
\if{latex}{\out{\hypertarget{method-plotIndiv}{}}}
\subsection{Method \code{plotIndiv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotIndiv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotArrow"></a>}}
\if{latex}{\out{\hypertarget{method-plotArrow}{}}}
\subsection{Method \code{plotArrow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotArrow()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVar"></a>}}
\if{latex}{\out{\hypertarget{method-plotVar}{}}}
\subsection{Method \code{plotVar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotVar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCircos"></a>}}
\if{latex}{\out{\hypertarget{method-plotCircos}{}}}
\subsection{Method \code{plotCircos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotCircos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-plotLoadings}{}}}
\subsection{Method \code{plotLoadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotLoadings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNetwork"></a>}}
\if{latex}{\out{\hypertarget{method-plotNetwork}{}}}
\subsection{Method \code{plotNetwork()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotNetwork(cutoff = 0.4, output = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotHeatmap}{}}}
\subsection{Method \code{plotHeatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$plotHeatmap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cytoscapeImportTable"></a>}}
\if{latex}{\out{\hypertarget{method-cytoscapeImportTable}{}}}
\subsection{Method \code{cytoscapeImportTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$cytoscapeImportTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixomicsClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
