% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{NgsProjectClass}
\alias{NgsProjectClass}
\title{R6 class for project information}
\value{

}
\description{
R6 class for project information

R6 class for project information
}
\examples{
project <- NgsProjectClass$new(projectdir)
getGroupLevels returns unique ordered levels for specified group


## ------------------------------------------------
## Method `NgsProjectClass$getGroupLevels`
## ------------------------------------------------

getGroupLevels(project, groupcol='default')
getSamplesByGroup


## ------------------------------------------------
## Method `NgsProjectClass$getSamplesByGroup`
## ------------------------------------------------


##################################################################

#' getCountsByGroup adjusts the counts and samples to exclude missing samples
#'
#' @param counts RNAseq count data
#' @param project project data containing sample and group information
#' @param groupcol group column
#'
#' @return
#' @export
#'
#' @examples
getCountsByGroup = function(counts, project, groupcol)
{
   # subset samples non-null for the selected groupcol
   samples <- hlsgr::getSamplesByGroup(project, groupcol)
   # susbet the count data to include only samples selected above
   counts <- counts[, colnames(counts) \%in\% samples$sample]
   # makes sure that the sample list also matches samples in the count data
   samples <- samples[samples$name \%in\% colnames(counts),]
   return(list(samples = samples, counts = counts, groupcol = groupcol))
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NgsProjectClass$new()}}
\item \href{#method-getGroupLevels}{\code{NgsProjectClass$getGroupLevels()}}
\item \href{#method-getSamplesByGroup}{\code{NgsProjectClass$getSamplesByGroup()}}
\item \href{#method-clone}{\code{NgsProjectClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$new(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{folder containing project and sample information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getGroupLevels(groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupcol}}{group column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{getGroupLevels(project, groupcol='default')
getSamplesByGroup

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSamplesByGroup"></a>}}
\if{latex}{\out{\hypertarget{method-getSamplesByGroup}{}}}
\subsection{Method \code{getSamplesByGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getSamplesByGroup(groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupcol}}{group column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
##################################################################

#' getCountsByGroup adjusts the counts and samples to exclude missing samples
#'
#' @param counts RNAseq count data
#' @param project project data containing sample and group information
#' @param groupcol group column
#'
#' @return
#' @export
#'
#' @examples
getCountsByGroup = function(counts, project, groupcol)
{
   # subset samples non-null for the selected groupcol
   samples <- hlsgr::getSamplesByGroup(project, groupcol)
   # susbet the count data to include only samples selected above
   counts <- counts[, colnames(counts) \%in\% samples$sample]
   # makes sure that the sample list also matches samples in the count data
   samples <- samples[samples$name \%in\% colnames(counts),]
   return(list(samples = samples, counts = counts, groupcol = groupcol))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
