% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{WgcnaClass}
\alias{WgcnaClass}
\title{create R6 class method for WGCNA}
\description{
create R6 class method for WGCNA

create R6 class method for WGCNA
}
\examples{
wgcna <- WgcnaClass$new(result, datTraits)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WgcnaClass$new()}}
\item \href{#method-plotSampleTree}{\code{WgcnaClass$plotSampleTree()}}
\item \href{#method-cutoffline}{\code{WgcnaClass$cutoffline()}}
\item \href{#method-cutTreeStatic}{\code{WgcnaClass$cutTreeStatic()}}
\item \href{#method-cutTreeDynamic}{\code{WgcnaClass$cutTreeDynamic()}}
\item \href{#method-plotDendogram}{\code{WgcnaClass$plotDendogram()}}
\item \href{#method-pickSoftThreshold}{\code{WgcnaClass$pickSoftThreshold()}}
\item \href{#method-plotSoftThreshold}{\code{WgcnaClass$plotSoftThreshold()}}
\item \href{#method-plotMeanConnectivity}{\code{WgcnaClass$plotMeanConnectivity()}}
\item \href{#method-constructNetwork}{\code{WgcnaClass$constructNetwork()}}
\item \href{#method-plotDendogramWithColors}{\code{WgcnaClass$plotDendogramWithColors()}}
\item \href{#method-plotTomPlot}{\code{WgcnaClass$plotTomPlot()}}
\item \href{#method-plotModuleTraitRelationship}{\code{WgcnaClass$plotModuleTraitRelationship()}}
\item \href{#method-selectHubs}{\code{WgcnaClass$selectHubs()}}
\item \href{#method-goEnrichmentAnalysis}{\code{WgcnaClass$goEnrichmentAnalysis()}}
\item \href{#method-exportCytoscape}{\code{WgcnaClass$exportCytoscape()}}
\item \href{#method-calcPvalue}{\code{WgcnaClass$calcPvalue()}}
\item \href{#method-clone}{\code{WgcnaClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$new(
  result,
  datTraits,
  degenes_only = FALSE,
  nGenes = 5000,
  genecode = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{DESeq2 results object}

\item{\code{datTraits}}{clinicaldata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSampleTree"></a>}}
\if{latex}{\out{\hypertarget{method-plotSampleTree}{}}}
\subsection{Method \code{plotSampleTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotSampleTree()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutoffline"></a>}}
\if{latex}{\out{\hypertarget{method-cutoffline}{}}}
\subsection{Method \code{cutoffline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$cutoffline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutTreeStatic"></a>}}
\if{latex}{\out{\hypertarget{method-cutTreeStatic}{}}}
\subsection{Method \code{cutTreeStatic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$cutTreeStatic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cutTreeDynamic"></a>}}
\if{latex}{\out{\hypertarget{method-cutTreeDynamic}{}}}
\subsection{Method \code{cutTreeDynamic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$cutTreeDynamic(deepSplit = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDendogram"></a>}}
\if{latex}{\out{\hypertarget{method-plotDendogram}{}}}
\subsection{Method \code{plotDendogram()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotDendogram(signed = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickSoftThreshold"></a>}}
\if{latex}{\out{\hypertarget{method-pickSoftThreshold}{}}}
\subsection{Method \code{pickSoftThreshold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$pickSoftThreshold()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSoftThreshold"></a>}}
\if{latex}{\out{\hypertarget{method-plotSoftThreshold}{}}}
\subsection{Method \code{plotSoftThreshold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotSoftThreshold()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMeanConnectivity"></a>}}
\if{latex}{\out{\hypertarget{method-plotMeanConnectivity}{}}}
\subsection{Method \code{plotMeanConnectivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotMeanConnectivity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-constructNetwork"></a>}}
\if{latex}{\out{\hypertarget{method-constructNetwork}{}}}
\subsection{Method \code{constructNetwork()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$constructNetwork(
  minModuleSize = 30,
  maxBlockSize = 5000,
  mergeCutHeight = 0.25,
  TOMType = "unsigned"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDendogramWithColors"></a>}}
\if{latex}{\out{\hypertarget{method-plotDendogramWithColors}{}}}
\subsection{Method \code{plotDendogramWithColors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotDendogramWithColors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotTomPlot"></a>}}
\if{latex}{\out{\hypertarget{method-plotTomPlot}{}}}
\subsection{Method \code{plotTomPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotTomPlot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotModuleTraitRelationship"></a>}}
\if{latex}{\out{\hypertarget{method-plotModuleTraitRelationship}{}}}
\subsection{Method \code{plotModuleTraitRelationship()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$plotModuleTraitRelationship()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectHubs"></a>}}
\if{latex}{\out{\hypertarget{method-selectHubs}{}}}
\subsection{Method \code{selectHubs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$selectHubs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-goEnrichmentAnalysis"></a>}}
\if{latex}{\out{\hypertarget{method-goEnrichmentAnalysis}{}}}
\subsection{Method \code{goEnrichmentAnalysis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$goEnrichmentAnalysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exportCytoscape"></a>}}
\if{latex}{\out{\hypertarget{method-exportCytoscape}{}}}
\subsection{Method \code{exportCytoscape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$exportCytoscape(threshold = 0.02)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcPvalue"></a>}}
\if{latex}{\out{\hypertarget{method-calcPvalue}{}}}
\subsection{Method \code{calcPvalue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$calcPvalue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WgcnaClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
